register_handlers fn to add handlers after init
The register_handlers function should allow an application to add extra
handlers to an already transit enabled application.

This could be used if an application doesn't have access to all it's
handlers at the time init_app is called, or could be used by transit
python plugins that need to provide their own initialization outside of
calling init_app.